refactor: Replace dyn with EventHandlerType enum
